Add a tweaked version of tby-sd1 convention
The convention provides imports, contexts, and overrides. We are
pretty close to tby-sd1 ("scientific dataset"), but some changes will
be needed.

The first change, done here, is to provide different keys to funding
context). This means that the dataset convention needs to change, as
the funding file name is now different.