Add "sfbHomepage" to compaction context
Things that are not corresponding to catalog's schema directly, but
are meant to land in "additional display" and have IRIs assigned
during expansion can probably be mapped into sfb* keys for further
use.

Ideally, we could generate a nested "additional_display" object and
map onto its keys, but as far as I understand, restructuring the
original object is not possible with expansion/compaction alone. Thus,
I am also beginning to understand the concerns about placing these
properties in the "dataset" table.