Add publications handling
Publications are now added to the convention.

Publications are handled by context expansion/compaction plus
rudimentary custom processing (which does not yet cover the case when
doi is given and citation is not). DOI lookups are not yet performed.

Note that context compaction is able to give us doi and publication
date "for free".