Handle used-for when adding to the catalog
This takes a simplistic approach: squeezes all "used for" items into a
single string and puts them in the sfb-specific properties table as
"used for" or "used for (n) if there are more items.

Other than making "used for" a separate tab, I don't see a good
alternative way to display this piece of information.