Generate UUIDs based on project and dataset names
This is an arbitrary convention: UUID will be generated using
"sfb151.project.datasetname" as input. Project name will be lowercase,
and in case of multiple projects declared, only the first one will be
used.

The basic procedure (UUID v5 with a datalad namespace) was borrowed
from a draft in our catalog-utilities repo; the particular template is
the novelty.