Add ontology search lookups for UBERON and NCBITaxon
This implements simple API queries and formatting of outputs for
Uberon and NCBITaxon terms.

Terms are looked up https://www.ebi.ac.uk/ols4 API, and translated
into catalog reader-facing strings by looking up specific
keys. Queries are cached, with requests_cache library, in
query_cache.sqlite file.

Functions that query the api and process the response were placed in a
separate file to not overburden the main script. That file, however,
is a bit of a mess at the moment, as it contains half-baked attempts
of querying crossref about DOIs, which aren't currently used.