Add a script for renaming of tsv files
This expands load_excel to also take a bunch of prefixed tsv files and
copy them into a "directory-of-files". This would not be necessary as
such, but I currently want to do this for two reasons. One is the
desire for unification, the other (more important) is that context
machinnery works better in this layout.