Add super/subdataset relations
If a tabby file (collection) is placed in dataset, the code will now
produce an entry for that dataset with a "subdataset" content. This
effectively makes tabby a subdataset declaration.

Currently we are making some assumptions about the subdataset path
that may need to be revised: we use the tabby-file parent, which would
work well for a file-in-a-directory.