Add a script for creating tabby-style file lists
The script creates a tabby tsv file with a listing of files in a git
worktree. This can be useful e.g. when describing a collection of
files that were published to github(-like), without going the
subdataset route (e.g. to avoid the need to determine a custom dataset
id for catalog purposes).