Add a superdataset configuration (subdataset tab default)
Add a dataset-level configuration file to be used for
project-superdatasets. It is a copy from the sfb 1451 catalog repo,
and notably contains dataset_options.default_tab = subdatasets.

This allows us to work with a default that shows a dataset with
content tab open (which is a preferred way to show project
subdatasets) while making an "exception" for the two top levels (sfb
and project) which would show subdatasets first - introduced in
https://github.com/psychoinformatics-de/sfb1451-projects-catalog/pull/64