Allow for dataset self-description using tabby
Tabby is a useful way of communicating descriptive metadata (of the
kind that we want to include in the catalog). This commit enables
dataset self-description.

A tabby file will be recognize as self-description if it is placed in
.datalad/tabby/self/ (or named .datalad/tabby/self*). In this case,
dataset id and version will be taken from the file's parent
dataset. All other processing is done the same way as for any other
tabby.

Since the tabby files used for self-description are unlikely to be the
only metadata source for a given dataset, the --remove-first option
will be ignored for such file (to avoid removing metadata from other
sources when we are regenerating tabby information).