Add an alternative file lister implementation
This adds a file listing implementation based on ls_file_collection
from datalad-next (as opposed to status from datalad-core).

Requires datalad-next 1.1.0

This is not necessarily faster, but can produce more complete results,
as it will calculate and report md5sum for non-annexed files.

Code-wise the structure remains the same. We still need to look up the
MD5 sum from the annex keys because iter_annextree will only checksum
files it can access. Here, we allow ourselves to assume that the annex
key will use the and not another algorithm - something that is not
necessarily the case.

To (lazily) avoid having to rebuild the catalogging environment, I
keep both implementations for now, although I don't expect much effort
being involved in updating datalad-next dependency.