Allow load_tabby to reuse homepage as clone URL
SFB tabby format currently has no concept of clone URL but in some
cases (e.g. gin) a homepage can be interpreted as such. This would
allow the catalog to display "Download with DataLad" & "View on ..."
buttons.

If the homepage looks like a cloneable URL (which in practice is
tested by checking whether it is GIN, GitHub or GitLab), it will be
added to the metadata as the dataset's URL.

Note that using tabby to represent GIN repos (instead of installing
them as subdatasets) helps us avoid the need to add the DataLad
dataset ID (and any other metadata) to the repo.