Make doi.org queries return the doi as URL
The doi.org queries for CSL metadata return the DOI as just the
identifier, without the https://doi.org/ part. This would not work
well with the catalog, which uses the doi provided as-is as the link
target.

This change makes sure we always produce the publication doi as URL.