Introduce lookup tables for SFB1451 grant subprojects
The tabby format specification which we circulated to the SFB members
says that the SFB grant identifier should be provided in the form
expected for use in publications, i.e. 431549029-<CRC-project-code>.

This is sufficient for us to build the URL identifier pointing to the
GEPRIS entry for the main project (by simply appending the number to a
constant URL base), but not for the subproject, which has its own
entry in GEPRIS. Since we now want the catalog to communicate both
identifiers (as two separate "grant" entries), this commit introduces
a lookup table for GEPRIS subproject identifiers. Subproject names are
also included in the lookup.

If no match is found for the subproject code provided in the tabby
file, then only one entry, referring to the whole SFB, is returned.