Save requests-cache files in the default user cache dir
The requests-cache files are now renamed to "tabby-utils-queries" and
saved in the default user cache dir (`~/.cache` on my system; see
output of `appdirs.user_cache_dir()` to check) instead of the working
directory. This should prevent having the cache files scattered all
around, while still allowing to find and remove the file if needed.