Add a directory listing utility script
This adds a script for generating a listing of directory files. It is
an attempt at creating a recursive generator from ls_file_collection
with type = directory (in which case, the command works
non-recursively). It is kept separate from list_files (which was meant
to be flexible) due to the recursion - might want to merge in the
future.

The implementation goes into directories and yields files. It ignores
any other type (likely: symlink, annexed file).