Ensure that the dataset doi is in a URL form
Although our example tabby record used the URL form in the doi field,
the specification at https://rdm.sfb1451.de/data-catalog-submission/
does not forbid having a plain DOI (without the https part).

As it stands, the catalog uses the doi field as-is, leading to a
nonsensical local link being shown when the doi is not a URL.

This simplified solution does not account for things like "doi: ..."
or "dx.doi.org/..." but it should be easy to extend (if starts with
neither 10 nor http, use some regex...), but hopefully these can also
be caught when we implement proper validation.

Side note on regexes:
https://www.crossref.org/blog/dois-and-matching-regular-expressions/