Add ability to retrieve grant titles from GEPRIS & CORDIS
CORDIS (European Comission) offers XML representation of grant pages, and
GEPRIS (DFG) website can be parsed. This allows us to present grants
other than SFB1451 with a similar level of detail.

For DFG, we rely on funder name to be entered as "DFG" (which we then
expand), and for EU we rely on the grant identifier being a CORDIS
URL (and we take funder as-is). This is slightly inconsequential and
can be changed in the future - but for now I'm not tweaking the code
for handling the SFB1451 grant entries.

Closes #18